+++ packages/snaps-controllers/src/snaps/SnapController.ts
@@ -389,14 +389,6 @@ export type SnapStateChange = {
 };
 
 /**
- * Emitted when a Snap has been added to state during installation.
- */
-export type SnapAdded = {
-  type: `${typeof controllerName}:snapAdded`;
-  payload: [snap: Snap, svgIcon: string | undefined];
-};
-
-/**
  * Emitted when an installed snap has been blocked.
  */
 export type SnapBlocked = {
@@ -418,15 +410,6 @@ export type SnapInstalled = {
  */
 export type SnapUninstalled = {
   type: `${typeof controllerName}:snapUninstalled`;
-  payload: [snap: TruncatedSnap];
-};
-
-/**
- * Emitted when a snap is removed from state, this may happen even
- * if a snap has not fully completed installation.
- */
-export type SnapRemoved = {
-  type: `${typeof controllerName}:snapRemoved`;
   payload: [snap: TruncatedSnap];
 };
 
@@ -481,11 +464,9 @@ export type SnapDisabled = {
 };
 
 export type SnapControllerEvents =
-  | SnapAdded
   | SnapBlocked
   | SnapInstalled
   | SnapUninstalled
-  | SnapRemoved
   | SnapStateChange
   | SnapUnblocked
   | SnapUpdated
@@ -1503,8 +1484,6 @@ export class SnapController extends BaseController<
           delete state.snapStates[snapId];
         });
 
-        this.messagingSystem.publish(`SnapController:snapRemoved`, truncated);
-
         // If the snap has been fully installed before, also emit snapUninstalled.
         if (snap.status !== SnapStatus.Installing) {
           this.messagingSystem.publish(
@@ -2348,21 +2327,12 @@ export class SnapController extends BaseController<
       }
     }
 
-    const stringifiedIcon = svgIcon?.toString();
-
-    // TODO: Consider removing this as it is unused now
-    this.messagingSystem.publish(
-      `SnapController:snapAdded`,
-      snap,
-      stringifiedIcon,
-    );
-
     this.messagingSystem.call('SubjectMetadataController:addSubjectMetadata', {
       subjectType: SubjectType.Snap,
       name: proposedName,
       origin: snap.id,
       version,
-      svgIcon: stringifiedIcon ?? null,
+      svgIcon: svgIcon?.toString() ?? null,
     });
 
     return { ...snap, sourceCode };